Add files via upload
Title: News Authenticity Detection with Naive Bayes

Description:

This Kaggle project focuses on classifying news articles as either authentic or fake using a Naive Bayes classifier. The dataset includes both real and fake news articles, with features such as title, text, subject, and date.

Steps Taken:

Data Exploration:

Explored and loaded the dataset, consisting of real and fake news articles.
Checked the structure and content of the data.
Data Preprocessing:

Labeled the real news as 1 and fake news as 0.
Merged and shuffled the data to ensure a balanced distribution.
Text Classification:

Split the dataset into training and testing sets.
Implemented a Naive Bayes classifier using a pipeline with CountVectorizer and MultinomialNB.
Trained the model on the training data.
Model Evaluation:

Made predictions on the test set.
Evaluated the model's performance using accuracy and a classification report.
Results:

The Naive Bayes model achieved an impressive accuracy of approximately 95% on the test data. The precision, recall, and F1-score for both authentic and fake news demonstrate the effectiveness of the classifier.